+++ NEWS.md
@@ -36,6 +36,9 @@
 * Add `allcombinations` function that returns a data frame created
   from all combinations of the passed vectors
   ([#3031](https://github.com/JuliaData/DataFrames.jl/pull/3031))
+* Add `resize!`, `keepat!`, `pop!`, `popfirst!`, and `popat!`,
+  make `deleteat!` signature more precise
+   ([#3047](https://github.com/JuliaData/DataFrames.jl/pull/3047))
 * New `multithreaded` argument allows disabling multithreading in
   `combine`, `select`, `select!`, `transform`, `transform!`, `subset` and `subset!`
   ([#3030](https://github.com/JuliaData/DataFrames.jl/pull/3030))